// cluster/authserver/authserver/web/package/src/apis/metav1/metav1.ts
export interface DeleteOptions {
    /**
     * UID is the objet's UID.
     *
     * @generated from protobuf field: string uid = 1;
     */
    uid: string;
    /**
     * Name is the object name that needs to be deleted. Either the name or the
     * UID must be used.
     *
     * @generated from protobuf field: string name = 2;
     */
    name: string;
}
// ...
class DeleteOptions$Type extends MessageType<DeleteOptions> {
    constructor() {
        super("octelium.api.main.meta.v1.DeleteOptions", [
            { no: 1, name: "uid", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<DeleteOptions>): DeleteOptions {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.uid = "";
        message.name = "";
        if (value !== undefined)
            reflectionMergePartial<DeleteOptions>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: DeleteOptions): DeleteOptions {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string uid */ 1:
                    message.uid = reader.string();
                    break;
                case /* string name */ 2:
                    message.name = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: DeleteOptions, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string uid = 1; */
        if (message.uid !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.uid);
        /* string name = 2; */
        if (message.name !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.name);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.meta.v1.DeleteOptions
 */
export const DeleteOptions = new DeleteOptions$Type();